Rollback a problematic solution to having variable-size soil layers
The issue is that multiple places in the code expect specific number, and specific meaning, of soil layers. For example:
SoilTransport::fq_NTransport() fails in line:
const double NO3_u = vq_SoilNO3_aq[i + 1];

CropModule::fc_OxygenDeficiency() expects that there are at least 3 soil layers, and that layers 0, 1 and 2 are 3 layers from the top 30 cm of the soil

As a result, this code crashes for soil profiles configured to have e.g. only 1 or 2 layers defined -- but code is still incorrect (e.g. CropModule::fc_OxygenDeficiency()) even if there are more layers, as long as first 3 layers aren't 10cm each.